Replaced sha form field with a field called path
The removed sha key would be used when calling Github's contents API in
order to fetch the contents of each of the selected files.
Unfortunatelly this is not supported.
Instead, the root path of the library within the repository (might be
the same as the repo root directory) should be provided.
In order to achieve that, the path form field is updated each time the
jsTree selection event is fired.