Updated checkGithubUpdates function
Made the function compatible with the getLastCommit function. However,
the current state of the checkGithubUpdates function does not support
checking the git branch of an already uploaded library before checking
for the last commit. Thus the master branch was used for all cases and a
TODO was added in order to check the function code again.
Note: this function is not actually used ever. We need to figure
out what we want it to return.