Added function for fetching code from github.
The getGithubRepoCode function handles the fetching of the code of a
folder (given a repo-branch combination). For every file in the folder,
getGithubFileCode function will be called in order to get the actual
content of the file. For every subfolder, the getGithubRepoCode will be
called recursively, until all the code is fetched.