Handle utf-8 encoding of text files properly (applies to both git & zip)
Instead of converting all files' contents to UTF-8 (which breaks
binaries, executables, etc) write the file, then identify the mime-type
and if dealing with a text file that is not encoded in UTF-8, convert it
to UTF-8 and write the file again.
Might take longer to execute, but this is the safest way of converting
only text files to UTF-8.