Removed function that changes files encoding on upload.
As we already discussed, all data should be stored in our
databases/filesystems as is. That is without changing the character encoding
of the files, or any other of their properties.
Thus, the function was removed, and the file encoding should be changed
only while fetching the files of a library.